+++ tests/testthat/test-carehomes-multistrain.R
@@ -93,10 +93,10 @@ test_that("Adding empty strains makes no difference", {
   p1 <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
   p2 <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                              strain_transmission = c(1, 0),
-                             model_super_infection = 1)
+                             cross_immunity = 0)
   p3 <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                              strain_transmission = c(1, 0.5),
-                             model_super_infection = 1)
+                             cross_immunity = 0)
   np <- 10
   mod1 <- carehomes$new(p1, 0, np, seed = 1L)
   mod2 <- carehomes$new(p2, 0, np, seed = 1L)
@@ -129,7 +129,7 @@ test_that("Seeding of second strain generates an epidemic", {
                             strain_seed_date =
                               sircovid_date(c(date_seeding, date_seeding_end)),
                             strain_seed_rate = c(n_seeded_new_strain_inf, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, 1, seed = 1L)
   info <- mod$info()
@@ -258,7 +258,7 @@ test_that("N_tot, is constant with second strain and waning immunity, while
                             strain_seed_date =
                               sircovid_date(c(date_seeding, date_seeding_end)),
                             strain_seed_rate = c(n_seeded_new_strain_inf, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, 1)
   info <- mod$info()
@@ -316,7 +316,7 @@ test_that("No-one in strains 3 or 4 if waning_rate is 1e6", {
                             strain_seed_date =
                               sircovid_date(c(date_seeding, date_seeding_end)),
                             strain_seed_rate = c(n_seeded_new_strain_inf, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, 1, seed = 2L)
   info <- mod$info()
@@ -365,7 +365,7 @@ test_that("prob_strain sums to 1", {
                             strain_seed_date =
                               sircovid_date(c(date_seeding, date_seeding_end)),
                             strain_seed_rate = c(n_seeded_new_strain_inf, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   mod <- carehomes$new(p, 0, np, seed = 1L)
   initial <- carehomes_initial(mod$info(), np, p)
   mod$set_state(initial$state, initial$step)
@@ -388,7 +388,7 @@ test_that("prob_strain sums to 1", {
                             strain_seed_date =
                               sircovid_date(c(date_seeding, date_seeding_end)),
                             strain_seed_rate = c(n_seeded_new_strain_inf, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   mod <- carehomes$new(p, 0, np, seed = 1L)
   initial <- carehomes_initial(mod$info(), np, p)
   mod$set_state(initial$state, initial$step)
@@ -413,7 +413,7 @@ test_that("No infection after seeding of second strain with 0 transmission", {
                             strain_seed_date =
                               sircovid_date(c(date_seeding, date_seeding_end)),
                             strain_seed_rate = c(n_seeded_new_strain_inf, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, 1, seed = 1L)
   info <- mod$info()
@@ -441,7 +441,7 @@ test_that("Everyone is infected when second strain transmission is large", {
                             strain_seed_date =
                               sircovid_date(date_seeding),
                             strain_seed_rate = c(n_seeded_new_strain_inf, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   ## set gamma_E to Inf so that seeded individuals move through each E stage
   ## in one step
@@ -475,7 +475,7 @@ test_that("No infection with either strain with perfect vaccine", {
                             rel_p_sympt = c(1, 1),
                             rel_p_hosp_if_sympt = c(1, 1),
                             waning_rate = 0.1,
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   mod <- carehomes$new(p, 0, 1)
   info <- mod$info()
 
@@ -509,7 +509,7 @@ test_that("different strains are equivalent", {
   set.seed(1)
   p <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                             strain_transmission = c(1, 1),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   np <- 10
   mod <- carehomes$new(p, 0, np, seed = 1L, n_threads = 10)
   end <- sircovid_date("2020-07-31") / p$dt
@@ -554,7 +554,7 @@ test_that("different strains are equivalent", {
 test_that("Swapping strains gives identical results with different index", {
   p <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                             strain_transmission = c(1, 1),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   np <- 1
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -626,7 +626,7 @@ test_that("Cannot calculate Rt for multistrain without correct inputs", {
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   np <- 3L
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -708,7 +708,7 @@ test_that("Cannot calculate Rt for multistrain without correct inputs", {
 test_that("wtmean_Rt works as expected", {
   p <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                             strain_transmission = c(1, 1),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   np <- 3L
   mod <- carehomes$new(p, 0, np, seed = 1L)
   initial <- carehomes_initial(mod$info(), np, p)
@@ -790,7 +790,7 @@ test_that("Can calculate Rt with an empty second variant ", {
   ## no seeding for second variant so noone infected with that one
   p <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                             strain_transmission = c(1, 1),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   np <- 3L
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -869,7 +869,7 @@ test_that("Can calculate Rt with a second less infectious variant", {
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   np <- 3L
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -933,7 +933,7 @@ test_that("Can calculate Rt with a second more infectious variant", {
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
 
   np <- 3L
@@ -999,7 +999,7 @@ test_that("Can calculate Rt with a second less letal variant", {
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
 
   np <- 3L
@@ -1067,7 +1067,7 @@ test_that("Can calculate Rt with a second variant with longer I_A", {
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
 
   np <- 3L
@@ -1135,7 +1135,7 @@ test_that("Can calculate Rt with a second variant with longer I_P", {
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
 
   np <- 3L
@@ -1203,7 +1203,7 @@ test_that("Can calculate Rt with a second variant with longer I_C_1", {
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
 
   np <- 3L
@@ -1268,7 +1268,7 @@ test_that("If prob_strain is NA then Rt is NA ", {
   ## no seeding for second variant so noone infected with that one
   p <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                             strain_transmission = c(1, 1),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   np <- 3L
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -1338,7 +1338,7 @@ test_that("calculate Rt with both second variant and vaccination", {
                             vaccine_progression_rate = numeric(3),
                             vaccine_schedule = vaccine_schedule,
                             vaccine_index_dose2 = 2L,
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   np <- 3L
   mod <- carehomes$new(p, 0, np, seed = 2L)
@@ -1440,10 +1440,10 @@ test_that("Relative gamma = 1 makes no difference", {
   p1 <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
   p2 <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                              strain_transmission = c(1, 1),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   p3 <- carehomes_parameters(sircovid_date("2020-02-07"), "england",
                              strain_transmission = c(1, 1),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   np <- 10
   mod1 <- carehomes$new(p1, 0, np, seed = 1L)
   mod2 <- carehomes$new(p2, 0, np, seed = 1L)
@@ -1479,7 +1479,7 @@ test_that("Lower rate variant has higher Rt", {
                               c(sircovid_date("2020-02-07"),
                                 sircovid_date("2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   np <- 3L
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -1509,7 +1509,7 @@ test_that("Lower rate variant has higher Rt", {
                               c(sircovid_date("2020-02-07"),
                                 sircovid_date("2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   np <- 3L
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -1552,7 +1552,7 @@ test_that("Stuck when gamma =  0", {
                               c(sircovid_date("2020-02-07"),
                                 sircovid_date("2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   p$gamma_P[] <- 0
 
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -1584,7 +1584,7 @@ test_that("Stuck when gamma =  0", {
                               c(sircovid_date("2020-02-07"),
                                 sircovid_date("2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   p$gamma_C_1[] <- 0
 
   mod <- carehomes$new(p, 0, np, seed = 1L)
@@ -1655,7 +1655,7 @@ test_that("Stuck when gamma =  0 for second strain", {
                               c(sircovid_date("2020-02-07"),
                                 sircovid_date("2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, np, seed = 1L)
 
@@ -1702,7 +1702,7 @@ test_that("Stuck when gamma =  0 for second strain", {
                               c(sircovid_date("2020-02-07"),
                                 sircovid_date("2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, np, seed = 1L)
   mod$set_state(initial$state, initial$step)
@@ -1725,7 +1725,7 @@ test_that("Stuck when gamma =  0 for second strain", {
                               c(sircovid_date("2020-02-07"),
                                 sircovid_date("2020-02-08")),
                             strain_seed_rate = c(10, 0),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, np, seed = 1L)
   mod$set_state(initial$state, initial$step)
@@ -1747,7 +1747,7 @@ test_that("No one is hospitalised, no-one recovers in edge case 2 - multi", {
                             strain_seed_rate = c(1, 0),
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   p$p_H_step <- 1
   p$psi_H[] <- 1
   p$p_G_D_step <- matrix(1)
@@ -1866,7 +1866,7 @@ test_that("G_D empty when p_G_D = 0", {
                             strain_seed_rate = c(1, 0),
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   p$p_G_D[] <- 0
   p$psi_G_D[] <- 0
   p$p_G_D_step[] <- 0
@@ -1892,7 +1892,7 @@ test_that("G_D strain 2 empty when p_G_D = c(1, 0)", {
                             strain_seed_rate = c(10, 0),
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, np, seed = 1L)
 
@@ -1921,7 +1921,7 @@ test_that("Can't move from S to E3/4", {
                             strain_seed_rate = c(10, 0),
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, np, seed = 1L)
 
@@ -1946,7 +1946,7 @@ test_that("Nobody in R2-R4 when strain_transmission = c(1, 0)", {
                             beta_value = 1e100,
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
 
   mod <- carehomes$new(p, 0, np, seed = 1L)
 
@@ -1973,7 +1973,7 @@ test_that("Can only move to S from R3 and R4 to S", {
                             waning_rate = 1 / 5,
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   ## Prevent anyone leaving S
   p$rel_susceptibility[] <- 0
 
@@ -2010,7 +2010,7 @@ test_that("Everyone in R3 and R4 when no waning and transmission high", {
                             strain_seed_rate = c(10, 0),
                             strain_seed_date =
                               sircovid_date(c("2020-02-07", "2020-02-08")),
-                            model_super_infection = 1)
+                            cross_immunity = 0)
   p$strain_transmission[] <- 1e8
 
   mod <- carehomes$new(p, 0, np, seed = 1L)